更新题库

+++ util/util.go
@@ -6,9 +6,12 @@ import (
 	"strings"
 	"bytes"
 	"os/exec"
+	"net"
 )
 
 var segmenter sego.Segmenter
+
+var Auto = false
 
 func init() {
 	segmenter.LoadDictionary("data/dictionary.txt")
@@ -27,20 +30,38 @@ func Split(str string) []string {
 	return sego.SegmentsToSlice(segments, true)
 }
 
-
 func RunWithAdb(args ...string) {
 
-	log.Printf("adb %s", strings.Join(args, " "))
+	if Auto {
 
-	var buffer bytes.Buffer
-	cmd := exec.Command("adb", args...)
-	cmd.Stdout = &buffer
-	cmd.Stderr = &buffer
-	err := cmd.Run()
-	if cmd.Process != nil {
-		cmd.Process.Kill()
+		log.Printf("adb %s\n", strings.Join(args, " "))
+
+		var buffer bytes.Buffer
+		cmd := exec.Command("adb", args...)
+		cmd.Stdout = &buffer
+		cmd.Stderr = &buffer
+		err := cmd.Run()
+		if cmd.Process != nil {
+			cmd.Process.Kill()
+		}
+		if err != nil {
+			log.Printf("adb %s: %s\n", strings.Join(args, " "), err)
+		}
+
 	}
-	if err != nil {
-		log.Fatalf("adb %s: %v", strings.Join(args, " "), err)
+}
+
+func HostIP() (ip string) {
+	addrs, _ := net.InterfaceAddrs()
+
+	for _, address := range addrs {
+		// 检查ip地址判断是否回环地址
+		if ipnet, ok := address.(*net.IPNet); ok && !ipnet.IP.IsLoopback() {
+			if ipnet.IP.To4() != nil {
+				return ipnet.IP.String()
+			}
+
+		}
 	}
+	return
 }
